[update: folder structure sockets and events]
