+++ client/src/components/chat/chatDetail.scss
@@ -16,7 +16,8 @@
     left: 0;
     right: 0;
     z-index: 1;
-    background: white;
+    padding: 0 2em;
+    background: $white;
 
     a {
       &:nth-of-type(2) {
@@ -61,6 +62,7 @@
     flex-wrap: wrap;
     justify-content: space-around;
     flex-direction: column;
+    margin-top: 16vh;
 
     ul {
       display: flex;
@@ -71,7 +73,7 @@
       align-items: flex-start;
       li {
         list-style: none;
-        margin: .5em 0;
+        margin: 0.5em 0;
 
         div {
           // width: fit-content;
@@ -161,7 +163,7 @@
         }
       }
 
-      button{
+      button {
         display: none;
       }
     }
